+++ src/_utils/keycodes.ts
@@ -1,5 +1,7 @@
 export const KEYCODES = {
   ESCAPE: 27,
+  ENTER: 13,
+  SPACEBAR: 32,
 }
 
 export default KEYCODES